0x13. C - More singly linked lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,21 +1,16 @@
 #include "lists.h"
-#include <stdio.h>
 
 /**
- * print_listint - function that prints all the elements of a listint_t list.
- * @h: pointer to the head of linked list.
+ * print_listint - function with one argument
+ * @h: pointer to struct
  *
- * Format: see example.
- * You are allowed to use printf.
- *
- * Return: all the elements of the list.
+ * Description: prints all the elements of a list
+ * Return: number of nodes
  */
-
 size_t print_listint(const listint_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
-	/* while we have a pointer to head */
 	while (h != NULL)
 	{
 		printf("%d\n", h->n);
